refactor(server): Moved JsonPointers to zally-core #560
