+++ server/src/main/java/de/zalando/zally/rule/DefaultContext.kt
@@ -3,7 +3,7 @@ package de.zalando.zally.rule
 import com.fasterxml.jackson.core.JsonPointer
 import de.zalando.zally.rule.api.Context
 import de.zalando.zally.rule.api.Violation
-import de.zalando.zally.util.ast.JsonPointers
+import de.zalando.zally.core.JsonPointers
 import de.zalando.zally.util.ast.ReverseAst
 import io.swagger.models.Swagger
 import io.swagger.v3.oas.models.OpenAPI
