+++ server/src/main/java/de/zalando/zally/rule/zalando/UseProblemJsonRule.kt
@@ -11,7 +11,7 @@ import de.zalando.zally.rule.api.Context
 import de.zalando.zally.rule.api.Rule
 import de.zalando.zally.rule.api.Severity
 import de.zalando.zally.rule.api.Violation
-import de.zalando.zally.util.ast.JsonPointers
+import de.zalando.zally.core.JsonPointers
 import io.swagger.v3.oas.models.responses.ApiResponse
 
 @Rule(
