+++ server/src/main/java/de/zalando/zally/rule/zally/NoUnusedDefinitionsRule.kt
@@ -6,7 +6,7 @@ import de.zalando.zally.rule.api.Check
 import de.zalando.zally.rule.api.Rule
 import de.zalando.zally.rule.api.Severity
 import de.zalando.zally.rule.api.Violation
-import de.zalando.zally.util.ast.JsonPointers
+import de.zalando.zally.core.JsonPointers
 
 @Rule(
     ruleSet = ZallyRuleSet::class,
