+++ server/zally-core/src/main/kotlin/de/zalando/zally/core/JsonPointerExtensions.kt
@@ -1,0 +1,2 @@
+package de.zalando.zally.core
+
